carrinho

+++ app/assets/javascripts/application.js
@@ -46,7 +46,7 @@ $( document ).on('turbolinks:load', function() {
   });
 });
 
-function adiciona_carrinho(produto) {
+function adiciona_produto_carrinho(produto) {
   var carrinho = []
   var carrinho_log = JSON.parse(localStorage.getItem('carrinho'));
 
@@ -60,20 +60,70 @@ function adiciona_carrinho(produto) {
   console.log(carrinho);
 }
 
-function remove_carrinho(produto) {
-  var carrinho = []
-  var carrinho_log = JSON.parse(localStorage.getItem('carrinho'));
+function remove_produto_carrinho(produto) {
+  var answer = confirm("Confirmar remoção?")
+  if (answer) {
+    var carrinho = []
+    var carrinho_log = JSON.parse(localStorage.getItem('carrinho'));
 
-  if (carrinho_log != null)
-    carrinho = carrinho_log
+    if (carrinho_log != null) carrinho = carrinho_log
 
-  carrinho.filter(prod => prod.id != produto.id);
-  
-  localStorage.setItem('carrinho', JSON.stringify(carrinho));
+    produtos = carrinho.filter(prod => prod.id == produto);
+    produtos.pop();
+    carrinho = carrinho.filter(prod => prod.id != produto).concat(produtos);
+    localStorage.setItem('carrinho', JSON.stringify(carrinho));
+  }
+  else {
+      //some code
+  }
+  montar_carrinho();
 }
 
-// $(document).ready(function(){
-//   $("#add-carrinho").click(function(){
-//     alert("The paragraph was clicked.");
-//   });
-// });
+function montar_carrinho() {
+  $(".carrinho-body").empty();
+  $(".carrinho-footer").empty();
+  var carrinho_log = JSON.parse(localStorage.getItem('carrinho'));
+
+  if (carrinho_log.length != 0)
+    $(".carrinho-footer").append(`
+      <button type="button" class="btn btn-default" data-dismiss="modal">Fechar</button>
+      <button type="button" class="btn btn-primary">Finalizar Compra</button>
+    `);
+  else
+    $(".carrinho-footer").append(`
+      <button type="button" class="btn btn-default" data-dismiss="modal">Fechar</button>
+    `);
+
+  var preco_total = 0;
+  carrinho_log.forEach(produto => {
+    var qtd = carrinho_log.filter(prod => prod.id == produto.id).length;
+    var produto_html = `
+      <div class="col-12">
+        <div class="carrinho-produto">
+          <div class="carrinho-produto-body">
+            <div class='carrinho-produto-img'>
+              <img alt="image" class="img-responsive" src="https://images-americanas.b2w.io/produtos/01/00/sku/43816/4/43816428_1SZ.jpg">
+            </div>
+            <div class='carrinho-produto-info'>
+              <h5 class="card-title">` + produto.nome + `</h5>
+              <p class="card-text"><b>Descrição:</b> ` + produto.descricao + `</p>
+              <p class="card-text"><b>Preço:</b> ` + produto.preco + `</p>
+              <p class="card-text"><b>Quantidade:</b> ` + qtd + `</p>
+              <a class="btn btn-primary" onclick="remove_produto_carrinho('`+ produto.id +`')">remover do carrinho</a>
+            </div>
+          </div>
+        </div>
+      </div>
+    `;
+    preco_total += parseFloat(produto.preco);
+    $(".carrinho-body").append(produto_html);
+  });
+
+  $(".carrinho-body").append("<p>Total: R$ " + preco_total +"</p>");
+}
+
+$(document).ready(function(){
+  $("#open-carrinho").click(function(){
+    montar_carrinho();
+  });
+});
